# Rewrite the softmax op