--- a/onnxruntime/core/providers/cpu/math/softmax.cc
+++ b/onnxruntime/core/providers/cpu/math/softmax.cc
@@ -1,6 +1,5 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
-
 #include "core/providers/cpu/math/softmax.h"
 
 #include "core/framework/op_kernel.h"
@@ -8,53 +7,24 @@
 #include "core/providers/common.h"
 #include "core/providers/cpu/math/softmax_shared.h"
 #include "core/util/math.h"
+#include "core/util/eigen_common_wrapper.h"
 
 namespace onnxruntime {
 
-template <>
-Status Softmax<float>::Compute(OpKernelContext* ctx) const {
-  auto ctx_internal = static_cast<OpKernelContextInternal*>(ctx);
-  concurrency::ThreadPool* tp = ctx_internal->GetOperatorThreadPool();
-
-  const auto* tensor_pointer = ctx->Input<Tensor>(0);
-  if (tensor_pointer == nullptr) return Status(common::ONNXRUNTIME, common::FAIL, "input count mismatch");
-  const Tensor& X = *tensor_pointer;
-  const TensorShape& input_shape{X.Shape()};
-
-  VLOGS(ctx->Logger(), 2) << "Input tensor shape: " << input_shape;
-
-  Tensor* Y = ctx->Output(0, input_shape);
-
-  const int64_t axis = HandleNegativeAxis(axis_, input_shape.NumDimensions());
-
-  size_t N = input_shape.SizeToDimension(axis);
-  size_t D = input_shape.SizeFromDimension(axis);
-
-  auto* Ydata = Y->template MutableData<float>();
-
-  std::vector<float> scale_(N);
-  std::vector<float> rowmax_(N);
-  std::vector<float> sum_multiplier_(D, 1.f);  // initialize all multiplier values to 1.0
-
-  const bool logarithmic = false;
-  auto status = SoftmaxCPU(N, D, X.template Data<float>(), Ydata,
-                           scale_.data(), sum_multiplier_.data(), logarithmic, rowmax_.data(), tp);
-
-  return status;
-}
-
-ONNX_CPU_OPERATOR_VERSIONED_KERNEL(
-    Softmax,
-    1,
-    10,
-    KernelDefBuilder().TypeConstraint("T", DataTypeImpl::GetTensorType<float>()),
-    Softmax<float>);
+ONNX_CPU_OPERATOR_VERSIONED_KERNEL(Softmax, 1, 10,
+                                   KernelDefBuilder().TypeConstraint("T", DataTypeImpl::GetTensorType<float>()),
+                                   Softmax<float, false>);
 
 // Opset 11 starts to support Neg Axis.
-ONNX_CPU_OPERATOR_KERNEL(
-    Softmax,
-    11,
-    KernelDefBuilder().TypeConstraint("T", DataTypeImpl::GetTensorType<float>()),
-    Softmax<float>);
+ONNX_CPU_OPERATOR_KERNEL(Softmax, 11, KernelDefBuilder().TypeConstraint("T", DataTypeImpl::GetTensorType<float>()),
+                         Softmax<float, false>);
+
+ONNX_CPU_OPERATOR_VERSIONED_KERNEL(LogSoftmax, 1, 10,
+                                   KernelDefBuilder().TypeConstraint("T", DataTypeImpl::GetTensorType<float>()),
+                                   Softmax<float, true>);
+
+// Opset 11 starts to support Neg Axis.
+ONNX_CPU_OPERATOR_KERNEL(LogSoftmax, 11, KernelDefBuilder().TypeConstraint("T", DataTypeImpl::GetTensorType<float>()),
+                         Softmax<float, true>);
 
 }  // namespace onnxruntime