--- a/onnxruntime/core/providers/cpu/math/softmax.cc
+++ b/onnxruntime/core/providers/cpu/math/softmax.cc
@@ -5,7 +5,6 @@
 #include "core/framework/op_kernel.h"
 #include "core/framework/op_kernel_context_internal.h"
 #include "core/providers/common.h"
-#include "core/providers/cpu/math/softmax_shared.h"
 #include "core/util/math.h"
 #include "core/util/eigen_common_wrapper.h"