--- a/onnxruntime/core/providers/cpu/math/softmax.h
+++ b/onnxruntime/core/providers/cpu/math/softmax.h
@@ -22,6 +22,7 @@ limitations under the License.
 #include "core/framework/op_kernel_context_internal.h"
 #include "core/util/math_cpuonly.h"
 #include "core/util/eigen_common_wrapper.h"
+#include "core/providers/common.h"
 
 namespace onnxruntime {
 // copied from tensorflow/core/kernels/softmax_op.cc