--- a/onnxruntime/core/providers/cpu/math/softmax.h
+++ b/onnxruntime/core/providers/cpu/math/softmax.h
@@ -19,6 +19,7 @@ limitations under the License.
 
 #include "core/common/common.h"
 #include "core/framework/op_kernel.h"
+#include "core/framework/op_kernel_context_internal.h"
 #include "core/util/math_cpuonly.h"
 #include "core/util/eigen_common_wrapper.h"