--- a/onnxruntime/test/providers/cpu/math/hardmax_test.cc
+++ b/onnxruntime/test/providers/cpu/math/hardmax_test.cc
@@ -1,7 +1,6 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-#include "core/providers/cpu/math/softmax_shared.h"
 #include "gmock/gmock.h"
 #include "gtest/gtest.h"
 #include "test/providers/provider_test_utils.h"