--- a/onnxruntime/test/providers/cpu/math/softmax_test.cc
+++ b/onnxruntime/test/providers/cpu/math/softmax_test.cc
@@ -192,46 +192,5 @@ TEST(SoftmaxOperator, InvalidAxis) {
           "-10 is not in valid range [-2,1]");
 }
 
-TEST(SoftmaxOperator, TestInputTooLarge) {
-  float* ignored = nullptr;
-  concurrency::ThreadPool tp("", 1);
-  // N > INT32_MAX
-  int64_t N = int64_t(INT32_MAX) + 1;
-  int64_t D = 1;
-  auto status = SoftmaxCPU(N, D, ignored, ignored, ignored, ignored, true, ignored, &tp);
-  EXPECT_EQ(status.Code(), common::INVALID_ARGUMENT);
-
-  // D > INT32_MAX
-  N = 1;
-  D = int64_t(INT32_MAX) + 1;
-  status = SoftmaxCPU(N, D, ignored, ignored, ignored, ignored, true, ignored, &tp);
-  EXPECT_EQ(status.Code(), common::INVALID_ARGUMENT);
-
-  // N * D > INT32_MAX
-  N = int64_t(INT32_MAX) / 2;
-  D = 3;
-  status = SoftmaxCPU(N, D, ignored, ignored, ignored, ignored, true, ignored, &tp);
-  EXPECT_EQ(status.Code(), common::INVALID_ARGUMENT);
-
-  /*
-    common::Status SoftmaxCPU(const int64_t N,
-                              const int64_t D,
-                              const float* Xdata,
-                              float* Ydata,
-                              float* scale,
-                              const float* sum_multiplier,
-                              bool logarithmic,
-                              float* rowmax)
-    {
-        // the Math functions SoftmaxCPU uses only support int32_t as input, so enforce that
-        if (N * D > INT32_MAX || N > INT32_MAX || D > INT32_MAX)
-        {
-            std::ostringstream ss;
-            ss << "SoftmaxCPU inputs N, D and N * D must be < " << INT32_MAX << ". N=" << N << ", D=" << D;
-            std::string msg = ss.str();
-
-            return Status(common::ONNXRUNTIME, common::INVALID_ARGUMENT, msg);
-        }*/
-}
 }  // namespace test
 }  // namespace onnxruntime