… onnxruntime/core/providers/cpu/math/logsoftmax.cc
… onnxruntime/core/providers/cpu/math/logsoftmax.h
… onnxruntime/core/providers/cpu/math/softmax_shared.cc
… onnxruntime/core/providers/cpu/math/softmax_shared.h