**Description**: 

Copy the tensorflow's softmax implementation to there. 

**Motivation and Context**
- Why is this change required? What problem does it solve?
Improve batch perf.

|data size   | old  |new   |
|---|---|---|
| 1x1000  | 0.006ms  | 0.008ms  |
| 1000x1000  | 2.5ms  | 1.5ms  |



The old one is not parallelized well.
The new one has much less code.

- If it fixes an open issue, please link to the issue here.